+++ mods/Woodcutting/mod/data.lua
@@ -4,3 +4,9 @@ require("prototypes.group")
 
 
 require("prototypes.entity.tree_entities")
+
+require("prototypes.item.plank_item")
+require("prototypes.recipe.plank_recipe")
+
+require("prototypes.technology.plank_technology")
+require("prototypes.technology.tree_technology")